+++ src/components/AvatarStack/AvatarStackMoreButton.tsx
@@ -1,39 +1,46 @@
 import React from 'react';
 
-import type {AvatarSize} from '../Avatar';
-import {Avatar, DEFAULT_AVATAR_SIZE} from '../Avatar';
+import {DEFAULT_AVATAR_SIZE} from '../Avatar';
 import {block} from '../utils/cn';
 
-import i18n from './i18n';
+import {AvatarStackMore} from './AvatarStackMore';
+import type {AvatarStackMoreButtonProps} from './types';
 
 const b = block('avatar-stack');
-
-export type AvatarStackMoreButtonProps = Pick<
-    React.HTMLProps<HTMLButtonElement>,
-    'className' | 'onClick' | 'aria-label'
-> & {
-    size?: AvatarSize;
-    count: number;
-};
 
 export const AvatarStackMoreButton = React.forwardRef<
     HTMLButtonElement,
     AvatarStackMoreButtonProps
->(({className, size = DEFAULT_AVATAR_SIZE, onClick, count, 'aria-label': ariaLabel}, ref) => {
-    return (
-        <button
-            ref={ref}
-            type="button"
-            className={b('more-button', {size}, className)}
-            onClick={onClick}
-        >
-            <Avatar
-                text={`+${count}`}
-                size={size}
-                aria-label={ariaLabel || i18n('more', {count})}
-            />
-        </button>
-    );
-});
+>(
+    (
+        {
+            className,
+            badgeClassName,
+            size = DEFAULT_AVATAR_SIZE,
+            onClick,
+            count,
+            'aria-label': ariaLabel,
+            borderColor,
+        },
+        ref,
+    ) => {
+        return (
+            <button
+                ref={ref}
+                type="button"
+                className={b('more-button', {size}, className)}
+                onClick={onClick}
+            >
+                <AvatarStackMore
+                    className={badgeClassName}
+                    size={size}
+                    count={count}
+                    aria-label={ariaLabel}
+                    borderColor={borderColor}
+                />
+            </button>
+        );
+    },
+);
 
 AvatarStackMoreButton.displayName = 'AvatarStack.MoreButton';