+++ src/components/AvatarStack/types.ts
@@ -41,3 +41,17 @@ export interface AvatarStackProps {
      */
     renderMore?: (options: {count: number}) => React.ReactElement;
 }
+
+export type AvatarStackMoreProps = Pick<
+    React.HTMLProps<HTMLDivElement>,
+    'className' | 'aria-label'
+> & {
+    count: number;
+    size?: AvatarSize;
+    borderColor?: string;
+};
+
+export type AvatarStackMoreButtonProps = Pick<React.HTMLProps<HTMLButtonElement>, 'onClick'> &
+    AvatarStackMoreProps & {
+        badgeClassName?: string;
+    };